Factor out OpenMP nested enabling/disabling
The enabling/disabling of OpenMP nested levels was done in a confusing
way in the KDTree constructors, with too much noise to make sane sense
of it. Moving this out into its own little class that does the correct
thing at destruction time makes things more readable on KDTree, plus it
removes the need for macro checks there.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>